backend/requirements.txt
# Core web framework
flask>=3.0.0
flask-cors>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# AI/ML APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Additional dependencies
werkzeug>=3.0.0

